% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{annotate_label_grey}
\alias{annotate_label_grey}
\title{Annotating plots with a grey background}
\usage{
annotate_label_grey(
  label,
  x,
  y,
  size = 4,
  fill = "#EBEBEB99",
  hjust = 0,
  vjust = 0,
  label.size = 0,
  ...
)
}
\arguments{
\item{label}{Text to write on the plot.}

\item{x, y}{x and y positions.}

\item{size, fill, hjust, vjust, label.size}{Plotting aesthetics that this function handles. They can be overridden.}

\item{...}{Other parameters to pass onto \code{\link[ggplot2:annotate]{ggplot2::annotate()}}\code{.}}
}
\value{
An annotation layer for a ggplot2 plot.
}
\description{
Annotating plots with a grey background
}
