% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{compare_pairs}
\alias{compare_pairs}
\title{Compare pairs of categorical variables}
\usage{
compare_pairs(data, levels, values, f = `-`)
}
\arguments{
\item{data}{a dataframe}

\item{levels}{a column with a categorical variable. All pairs of values in
\code{levels} will be compared.}

\item{values}{a column with values to compare.}

\item{f}{comparison function to apply to values in each pair. Defaults to \code{-}
to compute the pairwise differences.}
}
\value{
a dataframe with pairwise comparisons
}
\description{
Compare pairs of categorical variables
}
\examples{
to_compare <- nlme::Machines \%>\%
  dplyr::group_by(Worker) \%>\%
  dplyr::summarise(avg_score = mean(score)) \%>\%
  print()

to_compare \%>\%
  compare_pairs(Worker, avg_score) \%>\%
  dplyr::rename(difference = value) \%>\%
  dplyr::mutate_if(is.numeric, round, 1)
}
