% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{compare_sets}
\alias{compare_sets}
\title{Compare two vectors using R's set operations}
\usage{
compare_sets(x, y)
}
\arguments{
\item{x, y}{vectors to compare}
}
\value{
a list with \code{lengths} (the lengths of the other elements), \code{x}, \code{y},
\code{unique(x)}, \code{unique(y)}, \code{setequal(x, y)}, \code{setdiff(x, y)}, \code{setdiff(y, x)}, \code{intersect(x, y)}, \code{union(x, y)}.
}
\description{
Compare two vectors using R's set operations
}
\examples{
yours <- c(1, 2, 3, 4, 4)
mine <- c(3, 5, 6, 4)
compare_sets(yours, mine)
}
