% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{fct_glue_labels}
\alias{fct_glue_labels}
\alias{fct_add_counts}
\title{Format the labels of a factor}
\usage{
fct_glue_labels(xs, fmt = "{levels}", first_fmt = fmt)

fct_add_counts(xs, fmt = "{levels} ({counts})", first_fmt = fmt)
}
\arguments{
\item{xs}{a factor}

\item{fmt}{glue-style format to use. Defaults to \code{"{levels}"} for
\code{fct_glue_labels()} and \code{"{levels} ({counts})"} for \code{fct_add_counts()}.}

\item{first_fmt}{glue-style format to use for very first label. Defaults to
value of \code{fmt}.}
}
\value{
a factor with the labels updated
}
\description{
Format the labels of a factor
}
\details{
At this point, only the magic variables \code{"{levels}"} and
\code{"{counts}"} are available ". In principle, others could be defined.
\code{fct_add_counts()} is a special case of \code{fct_glue_labels()}.
}
