% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggmatplot}
\alias{ggmatplot}
\title{Plot columns of a matrix}
\usage{
ggmatplot(x, x_axis_column = NULL, n_colors = 6, unique_rows = TRUE)
}
\arguments{
\item{x}{A matrix.}

\item{x_axis_column}{Index (number) of the column to plot for the \emph{x}-axis.
Defaults to \code{NULL} in which case it uses row index (number) as the
\emph{x}-axis.}

\item{n_colors}{Number of colors to cycle through. Defaults to 6.}

\item{unique_rows}{Whether to work first take the unique rows of the matrix.
Defaults to \code{TRUE}.}
}
\value{
a ggplot2 plot.
}
\description{
Creates plots of matrices like \code{\link[graphics:matplot]{graphics::matplot()}} but uses ggplot2,
defaults to drawing lines, and can specify a column to use for the
\emph{x}-axis.
}
