% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggpreview}
\alias{ggpreview}
\title{Preview a file that would be created by \code{ggsave()}}
\usage{
ggpreview(..., device = "png")
}
\arguments{
\item{...}{options passed onto \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}

\item{device}{the file extention of the device to use. Defaults to \code{"png"}.}
}
\description{
This function saves a plot to a temporary file with \code{ggsave()} and opens the
temporary file in the system viewer. This function is useful for quickly
previewing how a plot will look when it is saved to a file.
}
