% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{is_same_as_last}
\alias{is_same_as_last}
\alias{replace_if_same_as_last}
\title{Check for locally repeating values}
\usage{
is_same_as_last(xs)

replace_if_same_as_last(xs, replacement = "")
}
\arguments{
\item{xs}{a vector}

\item{replacement}{a value used to replace a repeated value. Defaults to
\code{""}.}
}
\value{
\code{is_same_as_last()} returns TRUE when \code{xs[n]} the same as \code{xs[n-1]}.
}
\description{
Check for locally repeating values
}
\examples{
xs <- c("a", "a", "a", NA, "b", "b", "c", NA, NA)
is_same_as_last(xs)
replace_if_same_as_last(xs, "")
}
