% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{jekyll_create_rmd_draft}
\alias{jekyll_create_rmd_draft}
\title{Create a Jekyll draft post}
\usage{
jekyll_create_rmd_draft(
  slug = NULL,
  date = NULL,
  dir_drafts = "./_R/_drafts",
  open = TRUE
)
}
\arguments{
\item{slug}{A "slug" to use for the post. Should be a string consisting of
\verb{"hypen-separated-content-words}. Defaults to \code{NULL} in which case a random
slug is created.}

\item{date}{Date string to use for the post. Default to \code{NULL} for the
current date \code{format(Sys.Date())}.}

\item{dir_drafts}{Relative path to the folder to store the drafts. Defaults
to \verb{"./_R/_drafts}.}

\item{open}{Whether to open the file for editing when using RStudio. Defaults
to \code{TRUE}.}
}
\value{
The path to the created file is invisibly returned.
}
\description{
This is the function I use to create new posts for my website.
}
