% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{sample_n_of}
\alias{sample_n_of}
\title{Randomly sample data from n sub-groups of data}
\usage{
sample_n_of(data, size, ...)
}
\arguments{
\item{data}{a dataframe}

\item{size}{number of groups to sample}

\item{...}{variables to group by}
}
\value{
the data from subgroups
}
\description{
Randomly sample data from n sub-groups of data
}
\examples{
sample_data <- tibble::tibble(
  letter = rep(letters, 5),
  color = rep(c("red", "green", "yellow", "orange", "blue"), 26),
  value = rnorm(26 * 5)
)

# data from two letters
sample_data \%>\%
  sample_n_of(2, letter)

# data from two colors
sample_data \%>\%
  sample_n_of(2, color)

# data from 10 letter-colors pairs
sample_data \%>\%
  sample_n_of(10, letter, color)
}
