% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{seq_along_rows}
\alias{seq_along_rows}
\title{Create a sequence along the rows of a dataframe}
\usage{
seq_along_rows(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a sequence of integers along the rows of a dataframe
}
\description{
Create a sequence along the rows of a dataframe
}
