% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{str_which_between}
\alias{str_which_between}
\title{Which lines fall in between a delimeter pattern}
\usage{
str_which_between(string, pattern)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{a regular expression pattern to look for}
}
\value{
the lines that are contained between pairs of delimiter patterns
}
\description{
Which lines fall in between a delimeter pattern
}
\examples{
string <- "
```{r}
# some code
```

Here is more code.

```markdown
**bold!**
```
"

lines <- unlist(strsplit(string, "\n"))
str_which_between(lines, "^```")
}
