% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-stats.R
\name{tidy_correlation}
\alias{tidy_correlation}
\title{Generate tidy correlations}
\usage{
tidy_correlation(data, ..., type = c("pearson", "spearman"))
}
\arguments{
\item{data}{a dataframe}

\item{...}{columns to select, using \code{dplyr::select()} semantics.}

\item{type}{type of correlation, either \code{"pearson"} (the default) or
\code{"spearman"}.}
}
\value{
a long dataframe (a tibble) with correlations calculated for each
pair of columns.
}
\description{
This function respects groupings from \code{dplyr::group_by()}. When the dataframe
contains grouped data, the correlations are computed within each subgroup of
data.
}
\examples{
tidy_correlation(iris, -Species)

iris \%>\%
  dplyr::group_by(Species) \%>\%
  tidy_correlation(dplyr::starts_with("Petal"))
}
