% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-stats.R
\name{tidy_quantile}
\alias{tidy_quantile}
\title{Generate tidy quantiles for a dataframe column}
\usage{
tidy_quantile(data, var, probs = seq(0.1, 0.9, 0.2))
}
\arguments{
\item{data}{a dataframe}

\item{var}{a column in the dataframe}

\item{probs}{quantiles to return. Defaults to \code{c(.1, .3, .5, .7, .9)}}
}
\value{
a long dataframe (a tibble) with quantiles for the variable.
}
\description{
This function respects groupings from \code{dplyr::group_by()}. When the dataframe
contains grouped data, the quantiles are computed within each subgroup of
data.
}
\examples{
tidy_quantile(sleep, extra)

sleep \%>\%
  dplyr::group_by(group) \%>\%
  tidy_quantile(extra)
}
