% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjmisc-package.R
\docType{package}
\name{tjmisc-package}
\alias{tjmisc}
\alias{tjmisc-package}
\title{tjmisc: TJ's Miscellany}
\description{
A collection of helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/tjmisc}
  \item Report bugs at \url{https://github.com/tjmahr/tjmisc/issues}
}

}
\author{
\strong{Maintainer}: TJ Mahr \email{tristan.mahr@gmail.com}

}
\keyword{internal}
